requirements.txt
# Core Framework
chainlit>=1.0.0
streamlit>=1.28.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21

# Text Processing and NLP
langchain>=0.1.0
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
langdetect>=1.0.9
sentence-transformers>=2.2.2

# Vector Database
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Uncomment for GPU support

# Machine Learning
numpy>=1.24.0
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0

# HTTP and API
aiohttp>=3.8.0
requests>=2.31.0
httpx>=0.24.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utility Libraries
tqdm>=4.65.0
click>=8.1.0
rich>=13.0.0
typing-extensions>=4.7.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Logging and Monitoring
loguru>=0.7.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0  # For Excel file support

# Optional: Jupyter support for development
# jupyter>=1.0.0
# ipykernel>=6.25.0

# Optional: Additional document formats
# python-magic>=0.4.27  # File type detection
# ebooklib>=0.18  # EPUB support
# striprtf>=0.0.26  # RTF support

# Optional: Advanced text processing
# spacy>=3.6.0  # Advanced NLP
# nltk>=3.8.1  # Natural language toolkit

# Optional: Database support
# sqlalchemy>=2.0.0
# alembic>=1.12.0

# Optional: Async database
# asyncpg>=0.28.0  # PostgreSQL
# aiomysql>=0.2.0  # MySQL

# Optional: Caching
# redis>=4.6.0
# aioredis>=2.0.0

# Optional: Monitoring and metrics
# prometheus-client>=0.17.0
# structlog>=23.1.0

# Security
cryptography>=41.0.0
passlib>=1.7.4

# File handling
pathlib2>=2.3.7
watchdog>=3.0.0